# openapi.yaml
openapi: 3.1.0
info:
  title: Kancelaria API
  description: API for managing law firms and clients.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8000/api/v1
    description: Local development server
...
paths:
  /users/register:
    post:
      summary: Register a new user
      operationId: register_user
      tags:
        - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
        required: true
      responses:
        "201":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInDB"
        "400":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /users/token:
    post:
      summary: Authenticate user and get access token
      operationId: login_for_access_token
      tags:
        - users
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "401":
          description: Incorrect username or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /users/me:
    get:
      summary: Get current user
      operationId: read_users_me
      tags:
        - users
      security:
        - OAuth2PasswordBearer: []
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInDB"
        "401":
          description: Could not validate credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
        "400":
          description: Inactive user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /kancelarie:
    get:
      summary: Retrieve all law firms
      operationId: read_kancelarie
      tags:
        - kancelarie
      security:
        - OAuth2PasswordBearer: []
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KancelariaInDB"
    post:
      summary: Create a new law firm
      operationId: create_kancelaria
      tags:
        - kancelarie
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KancelariaCreate"
        required: true
      responses:
        "201":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KancelariaInDB"
...
components:
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          tokenUrl: /api/v1/users/token
          scopes: {}
  schemas:
    KancelariaBase:
      title: KancelariaBase
      type: object
      properties:
        name:
          type: string
          example: Law Firm A
        location:
          type: string
          example: Warsaw, Poland
        specialization:
          type: string
          example: Criminal Law
      required:
        - name
        - location
        - specialization
    KancelariaCreate:
      title: KancelariaCreate
      allOf:
        - $ref: "#/components/schemas/KancelariaBase"
...
    KancelariaInDB:
      title: KancelariaInDB
      allOf:
        - $ref: "#/components/schemas/KancelariaBase"
        - type: object
          properties:
            id:
              type: integer
              example: 1
          required:
            - id
    UserBase:
      title: UserBase
      type: object
      properties:
        email:
          type: string
          example: user@example.com
      required:
        - email
    UserCreate:
      title: UserCreate
      allOf:
        - $ref: "#/components/schemas/UserBase"
        - type: object
          properties:
            password:
              type: string
              example: securepassword
          required:
            - password
    UserInDB:
      title: UserInDB
      allOf:
        - $ref: "#/components/schemas/UserBase"
        - type: object
          properties:
            id:
              type: integer
              example: 1
            is_active:
              type: boolean
              default: true
              example: true
          required:
            - id
            - is_active
    Token:
      title: Token
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: bearer
      required:
        - access_token
        - token_type
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
    ValidationError:
      title: ValidationError
      type: object
      properties:
        loc:
          type: array
          items:
            type: string
        msg:
          type: string
        type:
          type: string
      required:
        - loc
        - msg
        - type